Write node data into individual files instead of in one big one.
